basic-multiagent/requirements.txt
python-dotenv=1.0.1
crewai=0.14.3